Added CSV processing functions
CSV is read and transformed in a vector of vector of strings named content.
Content variable is then processed to separate the parameters which are all ints
from the performance which is the result we will try to predict. printStringCSV
and printParamsAndResult are their to compare the two methods of printing,
from the string or from the processed values. Also added in datasets test.csv which is a very small csv file with the same structure.