Added Node struct, tree decision functions
Added struct Node that will serve to build up the decision tree. Added different mathematical  to get the max, min and mean of a list of parameters(this will be useful for deciding the threshhold).
Added random sampling function to get random rows that can serve as sample to determine the treshhold (instead of calculating on the whole dataset). Implemented bestThreshHoldColumn that determines the
best threshhold for a given value list of a specific feature. The decision is made on the minimization of the weighted variance. Created threshhold struct to hold the information outputed by
bestThreshHold Column.

Next step will be to start testing the functions to see how well they perform (O(n^2) complexity on  bestThreshholdColumn). Combine random sampling to find the best threshhold for each feature
and then the best feature on which to threshhold.
Nicolas