Added functions to find best threshold to split according to weighted variance.
findBestSplitRandom iterates through columns and for a given sample, find the best threshold in each column then takes the best feature to split on according to weighted variance. Started working on the
tree structure, nodeInitiate creates the first node but still a work in progress. The data split in node initiate is done with an index that is generated by checking which values are under or
over/equal to threshhold. This generates a vector<int> with 1 and 0s that tells how to split the data. The next task will be to tackle the recursiveness of the tree structure and allow it to generate
itself.
Also added printCSVLimited that only reads a n number of lines from the csv, which is useful.
Created the typedef Matrix from vector<vector<int>>.