Fixed nodeBuilder + clearer treePrint function + changed function parameters to references
nodeBuilder now works, node is handled as a pointer object and program can now create a tree. modified the node struct to include adress, which is a vector of 0s and 1s that indicates
which branch is where in the tree. Eg: 1111 is the right most leaf. also added data_size to Node struct, says how many rows/values go through this node. These new variables are useful
in the new treePrinter function and allows for a much clearer understanding of the tree.