Addition of the test_io.cpp program + addition of a comment in functions_io.cpp
The test_io.cpp program tests all the functions in functions_io.cpp, with the exception of ‘printStringCSV’ and ‘printParamAndResults’,
both of which are not very useful to test. There are also the ‘nodePrinter’ and ‘treePrinter’ functions,
which I wasn't able to test because there are errors in their implementations, and which may not be useful to test.