Improved implementation of the getMaxFeature, getMinFeature and getMeanFeature functions in the function_tree.cpp program + Added tests for certain functions in the function_tree.cpp program.
The changes to these functions are minor, they still work in the same way, I've just added conditions for cases where the matrix received as a parameter is empty, or if the index received as a parameter is out of range.
As far as the tests are concerned, they all pass correctly.