Modifying boosting.cpp, boosting.h, main.cpp and main_comparison to merge correctly without conflict
For boosting.cpp and .h, it was simple: all we had to do was join the small changes in the respective versions.

For main.cpp, it took me 5 hours (non-stop), and the changes included :
- Understanding the process behind the ./decision_tree executable
- Fixing certain parts of model_comparison, in particular the launch of the command that was executing MainEnsemble with several arguments (=+1 hours only on this point)
- Apply the ability to load a file by modifying main.cpp (in each method...) and main_comparaison.cpp.
- Fix various things that have been changed in main.cpp in the meantime (modified default values, default values displayed incorrectly, and other things... lots of other things...)
- Apply the ability to choose between MAE and MSE (change to main.cpp and main_comparison)

What's missing is the ability to choose between MAE and MSE, which I'm going to implement, but first finish the merge.