Continuing to linearise
Linearized boosting.cpp, bagging.cpp, boosting_XGBoost.cpp and decision_tree_XGBoost
Every change comes with a tests_ .cpp file to try out on sample data.
It converges for boosting and bagging, but for XGBosst it seems to converge not enough.
This may be due to the data set being very small, but at least there is a working version, try to find out how to make it converge better (dataset issue probably)