General reorganisation of the entire project + adjustment of includes in several files
A src/ file has been added to contain all the .cpp and .h files (with the main files) apart from the tests, which are still in the tests/ folder. In datasets two folders have been added, a raw/ with all the raw data and a processed/ with all the data modified by data_clean. The results folder now has 3 folders:
- a comparison/ containing the output from main_comparison.cpp.
- a saved_models/ containing all the models that have been saved by the save functions implemented in each bagging, boosting, ... file.
- a tree_visualization/ containing the visualizations (image in .png format) generated by tree_visualization.