Multithreading for singleDEcisioTree
Dependign on number of threads, the tree will be split until a certain depth.
For the moment, only a log2 number of threads is supported, with adjustNumThreads.
Added utility.h and .cpp to lighten main_ .cpp files. There currently is only one function in it but will refactor later.
Changed median calculation with a new class, much faster.